Make dateAgo work with correct locale setting
Added german translation

I couldn't get dateAgo to work with the correct language specific
strings if a local is set via MODx system settings, had to change
several things, but now it seems to work. Don't know if you want to
implement the changes =), but if, here they are!

I'm also unsure about the declension function, does it have any purpose
out of russian language? What is it for anyway?